+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Improved the flexibility of the WhoisServerList tests.
+
+### Added
+- Create an AbstractLocator and DomainLocator for more flexibility!
 
 ## [0.1.5] - 2018-05-01
 ### Removed